requirements.txt
aiofiles==0.7.0
aiopg==1.3.1
alembic==1.7.1
async-timeout==3.0.1
asyncpg==0.24.0
databases==0.5.1
environs==9.3.3
greenlet==1.1.1
httptools==0.3.0
importlib-resources==5.2.2
Jinja2==3.0.1
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow==3.13.0
multidict==5.1.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
pydantic==1.8.2
python-dotenv==0.19.0
sanic==21.6.2
sanic-envconfig==1.3.4
sanic-jinja2==0.10.0
sanic-routing==0.7.1
secure==0.3.0
SQLAlchemy==1.4.23
typing-extensions==3.10.0.2
websockets==9.1
zipp==3.5.0
